Pass mypy typechecking
